Alteração do local de inclusão de swagger do program.cs para o arquivo de extensão do servicecollection.
