+++ src/FiapCloudGames.Infrastructure/FiapCloudGames.Infrastructure.csproj
@@ -10,6 +10,8 @@
     <PackageReference Include="MongoDB.EntityFrameworkCore" Version="10.0.1" />
     <PackageReference Include="BCrypt.Net-Next" Version="4.1.0" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="10.0.5" />
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="10.0.5" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="10.1.7" />
   </ItemGroup>
 
   <ItemGroup>
